refactor: split types into a separate file and break down solveTCr function into private subfunctions
